finished refactoring home component
